// server/enums/strengthType.ts
enum StrengthType {
  // Broad/category level Strength Types
  SENSORY = 'SENSORY',
  EMOTIONS_FEELINGS_DEFAULT = 'EMOTIONS_FEELINGS_DEFAULT',
  PHYSICAL_SKILLS_DEFAULT = 'PHYSICAL_SKILLS_DEFAULT',
  LITERACY_SKILLS_DEFAULT = 'LITERACY_SKILLS_DEFAULT',
  NUMERACY_SKILLS_DEFAULT = 'NUMERACY_SKILLS_DEFAULT',
  ATTENTION_ORGANISING_TIME_DEFAULT = 'ATTENTION_ORGANISING_TIME_DEFAULT',
  LANGUAGE_COMM_SKILLS_DEFAULT = 'LANGUAGE_COMM_SKILLS_DEFAULT',
  PROCESSING_SPEED = 'PROCESSING_SPEED',
  MEMORY = 'MEMORY',

  // More specific Strength Types
  HANDWRITING = 'HANDWRITING',
  TURN_TAKING = 'TURN_TAKING',
  READING = 'READING',
  SPEED_OF_CALCULATION = 'SPEED_OF_CALCULATION',
  RESTFULNESS = 'RESTFULNESS',
  AUDITORY_DISCRIMINATION = 'AUDITORY_DISCRIMINATION',
  FOCUSING = 'FOCUSING',
  ARITHMETIC = 'ARITHMETIC',
  SPELLING = 'SPELLING',
  READING_EMOTIONS = 'READING_EMOTIONS',
  LANGUAGE_FLUENCY = 'LANGUAGE_FLUENCY',
  TIDINESS = 'TIDINESS',
  BALANCE = 'BALANCE',
  CREATIVITY = 'CREATIVITY',
  FINE_MOTOR_SKILLS = 'FINE_MOTOR_SKILLS',
  EMPATHY = 'EMPATHY',
  ESTIMATION = 'ESTIMATION',
  WRITING = 'WRITING',
  FINISHING_TASKS = 'FINISHING_TASKS',
  SOCIAL_ADAPTABILITY = 'SOCIAL_ADAPTABILITY',
  LISTENING = 'LISTENING',
  COMMUNICATION = 'COMMUNICATION',
  IMPULSE_CONTROL = 'IMPULSE_CONTROL',
  MATHS_LITERACY = 'MATHS_LITERACY',
  LEARNING_NEW_SKILLS = 'LEARNING_NEW_SKILLS',
  SENSORY_PROCESSING = 'SENSORY_PROCESSING',
  ALPHABET_ORDERING = 'ALPHABET_ORDERING',
  PROBLEM_SOLVING = 'PROBLEM_SOLVING',
  TASK_INITIATION = 'TASK_INITIATION',
  EXTROVERSION_INTROVERSION = 'EXTROVERSION_INTROVERSION',
  READING_COMPREHENSION = 'READING_COMPREHENSION',
  VISUAL_SKILLS = 'VISUAL_SKILLS',
  EMOTIONAL_CONTROL = 'EMOTIONAL_CONTROL',
  MATHS_CONFIDENCE = 'MATHS_CONFIDENCE',
  SPORTING_BALL_SKILLS = 'SPORTING_BALL_SKILLS',
  VISUAL_SPATIAL_SKILLS = 'VISUAL_SPATIAL_SKILLS',
  READING_VISUAL_DISCRIMINATION = 'READING_VISUAL_DISCRIMINATION',
  FRACTIONS_PERCENTAGES = 'FRACTIONS_PERCENTAGES',
  TIME_ALLOCATION = 'TIME_ALLOCATION',
  SOCIAL_NUANCES = 'SOCIAL_NUANCES',
  MANAGING_CHANGE = 'MANAGING_CHANGE',
  DUAL_TASKING = 'DUAL_TASKING',
  CONFIDENCE = 'CONFIDENCE',
  TRACKING = 'TRACKING',
  FOLDING_PACKING_SORTING = 'FOLDING_PACKING_SORTING',
  SPEAKING = 'SPEAKING',
  SELF_ORGANISED = 'SELF_ORGANISED',
  WORD_BASED_PROBLEMS = 'WORD_BASED_PROBLEMS',
  MONEY_MANAGEMENT = 'MONEY_MANAGEMENT',
  FORWARD_PLANNING = 'FORWARD_PLANNING',
  LANGUAGE_DECODING = 'LANGUAGE_DECODING',
  WORD_FINDING = 'WORD_FINDING',
  CALM = 'CALM',
  STAMINA = 'STAMINA',
  ACTIVE_LISTENING = 'ACTIVE_LISTENING',
  GRASP = 'GRASP',
  NUMBER_RECALL = 'NUMBER_RECALL',
  ATTENTION_TO_DETAIL = 'ATTENTION_TO_DETAIL',
  SPATIAL_AWARENESS = 'SPATIAL_AWARENESS',
  NON_VERBAL_COMMUNICATION = 'NON_VERBAL_COMMUNICATION',
  TASK_SWITCHING = 'TASK_SWITCHING',
  NUMBER_SEQUENCING = 'NUMBER_SEQUENCING',
  PEOPLE_PERSON = 'PEOPLE_PERSON',

  NONE = 'NONE',
}

export default StrengthType